+++ src/admins/admin.service.ts
@@ -8,7 +8,7 @@ import { AdminDto } from './dto/admin.dto';
 import { Admin } from './entities/admin.entity';
 import { AdminAuthorityRepository } from './repositories/adminAuthority.repository';
 import { EntityManager, getConnection } from 'typeorm';
-import { PostComplaints } from 'src/posts/postComplaints.entity';
+import { PostComplaints } from 'src/posts/entities/postComplaints.entity';
 import { PostComplaintsRepository } from 'src/posts/repositories/postComplaint.repository';
 import { SearchComplaintDto } from './dto/searchComplaint.dto';
 import { ChatComplaints } from 'src/chats/chatComplaints.entity';
