+++ src/posts/entities/complaintReason.entity.ts
@@ -2,7 +2,7 @@ import { UserComplaints } from '../../chats/userComplaints.entity';
 import { ChatComplaints } from '../../chats/chatComplaints.entity';
 import { Field, ObjectType } from '@nestjs/graphql';
 import { BaseEntity, Column, Entity, PrimaryGeneratedColumn, OneToMany } from 'typeorm';
-import { PostComplaints } from 'src/posts/postComplaints.entity';
+import { PostComplaints } from 'src/posts/entities/postComplaints.entity';
 
 @Entity()
 @ObjectType()
