+++ src/posts/entities/postsViewRecord.entity.ts
