+++ src/posts/entities/priceOffer.entity.ts
