+++ src/posts/repositories/post.repository.ts
@@ -10,7 +10,7 @@ import { UpdatePostDto } from '../dto/updatePost.dto';
 import { Post } from '../entities/post.entity';
 import { PriceOffer } from '../priceOffer.entity';
 import { ProcessState } from 'src/posts/entities/processState.entity';
-import { PostComplaints } from '../postComplaints.entity';
+import { PostComplaints } from '../entities/postComplaints.entity';
 import { DealState } from 'src/posts/entities/dealState.entity';
 import { User } from 'src/users/entities/user.entity';
 import { PostsLikeRecord } from '../postsLikeRecord.entity';
